package-indexer: move common logics to base classes
